+++ src/main/groovy/ru/vyarus/gradle/plugin/quality/QualityPlugin.groovy
@@ -209,6 +209,9 @@ class QualityPlugin implements Plugin<Project> {
                             withMessages true
                         }
                     }
+                    // in gradle 5 default 1g was changed and so spotbugs fails on large projects (recover behaviour),
+                    // but not if value set manually
+                    maxHeapSize = maxHeapSize ?: '1g'
                 }
             }
             configurePluginTasks(project, extension, SpotBugsTask, 'spotbugs', new SpotbugsReporter(configLoader))